add link generation code; testing pending
