+++ server/package.json
@@ -14,6 +14,7 @@
     "firebase-admin": "^11.5.0",
     "mongoose": "^7.0.0",
     "morgan": "^1.10.0",
+    "nodemailer": "^6.9.1",
     "uid": "^2.0.1"
   },
   "devDependencies": {
